feat:スキルが機能することを確認。魔素の表示位置を調整

+++ src/main/java/com/github/sweetfish111/reincarnated/system/CausalityObserver.java
@@ -23,7 +23,7 @@ public class CausalityObserver {
         if (event.getEntity() instanceof ServerPlayer player) {
             int xpAmount = event.getOrb().getValue();
 
-            Map<String, Object> data = Map.of("xp_amount", (float) xpAmount);
+            Map<String, Object> data = Map.of("xp_amount", (double) xpAmount);
             ActiveMagicManager.executeEventTrigger(player, EditorTab.SKILL, "on_xp_pickup", data);
 
             // プレイヤーの魂データを取り出して「貪欲者」としての理を進行させる
